Fixed setting of length in SQLGetData
Summary:
We used max_length which is maximum length of all columns in the rowset.
For NO_CACHE this value is equal to 0 as we are not fetching all rows.
And also it is incorrectly as we needed to return the actual column length instead of maximum.

Test Plan: https://app.circleci.com/pipelines/github/memsql/singlestore-odbc-connector/1974/workflows/0ea97cab-9a47-4632-85f9-1788597b5578

Reviewers: mshcherbina-ua, pmishchenko-ua

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D53159