Use lower xcode version
Summary: In order be able to use ODBC driver on MacOS 10 we need to use Xcode 11 or 12. Xcode provides forward compatibility, so the library compiled with Xcode 11 will be usable on systems with Xcode 12 and 13. In addition, libiodbc is added as a dependency into our installation package.

Test Plan: TBD

Reviewers: amakarovych-ua

Subscribers: engineering-list

JIRA Issues: PLAT-6130

Differential Revision: https://grizzly.internal.memcompute.com/D55623